ai/superseded/footy-legacy/requirements.txt
requests>=2.25.1
google-generativeai>=0.3.0
pathlib
opencv-python>=4.5.0
ffmpeg-python>=0.2.0
python-dotenv>=1.0.0 